+++ dataclass_array/utils/np_utils_test.py
@@ -23,7 +23,7 @@ import numpy as np
 import pytest
 
 # Activate the fixture
-enable_torch_tf_np_mode = enp.testing.enable_torch_tf_np_mode
+enable_tf_np_mode = enp.testing.set_tnp
 
 
 @enp.testing.parametrize_xnp()