+++ CHANGELOG.md
@@ -26,6 +26,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 *   **Add `torch` support!**
 *   Add `.cpu()`, `.cuda()`, `.to()` methods to move the dataclass from
     devices when using torch.
+*   **Breaking**: `@dataclass(frozen=True)` is now automatically applied
 
 ## [1.3.0] - 2023-01-16